feat: v1.3.1 智能识别增强 + 首页优化
- C1: VoiceCategorizer → SmartParseUseCase，三个入口统一调用
- 公司名自动识别为可报销
- 1万元正确解析为10000元
- 备注直接使用原始输入
- 统计界面报销明细跳转
- ¥符号修复
- 首页固定总览卡、可拖动FAB、日期折叠
- 设置界面顶格显示
- 标记报销自动创建收入记录
- 滑动冲突修复
- 记账模式适配
- Tab切换重置二级页面

+++ app/src/main/java/com/jiyixia/app/data/entity/Entities.kt
@@ -33,7 +33,7 @@ data class Category(
 data class Record(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val type: Int,              // 0=支出, 1=收入
-    val amount: Double,
+    val amount: Long,           // 金额，单位：分（38.10元 → 3810L）
     val categoryId: Long,
     val note: String = "",
     val date: Long,             // Unix 时间戳(毫秒)
